Added 'backup' in app/chatbot.dist.json
Changed backup() in chatbot/database.js
Changed README.md
Changed @babel/core 7.12.7 to 7.12.9
Changed vanilla-lazyload 17.1.3 to 17.3.0

+++ src/js/chatbot/database.js
@@ -1,3 +1,4 @@
+const config  = require('../../app/chatbot.json');
 const locales = require('./locales');
 const fs      = require('fs');
 const moment  = require('moment');
@@ -29,8 +30,8 @@ const database = {
         let backupFile = 'chatbot.' + moment().format(format) + '.sqlite3';
         let backupFolder = './data/backup/';
 
-        // if database exists and backup databse not exists
-        if (fs.existsSync(this.file) && !fs.existsSync(backupFolder + backupFile)) {
+        // if database exists and backup databse not exists and backup is enabled
+        if (fs.existsSync(this.file) && !fs.existsSync(backupFolder + backupFile) && (typeof config.backup === 'boolean' && config.backup)) {
             fs.copyFile(this.file, backupFolder + backupFile, (err) => {
                 if (err) {
                     throw err;
